feat(preact): register agnostic JSX host via jsx-runtime module
Register the host (jsx/jsxs/Fragment + childrenOf/typeOf/propsOf/isElement)
through the shared @expressive/mvc/jsx-runtime seam. Own shared proto seams in
runtime (rename intercept->ignore), inline Component.on(prepare), and align
runtime/component module names with their contents.
